Stub code is ready... I think
I decided to add a sort class for simplicity and so we don’t jam the
driver. The stub code should be pretty self explanatory. Feel free to
add whatever you feel necessary.
I have a pretty good idea of how to do all the sorting and organizing
of the data, so I will take that on shortly. Stanton, this should be
ready for you to do some javadoc.