First commit with file reader
This is a initial backup of my work so far. I have yet to test this yet
so don’t attempt to run it. What I have so far SHOULD read in and
assign objects based off the countries and continents data files. I
will do the cities file shortly. After this, I will begin working on
the different sort methods.