Took Awhile For this
Don’t know why i couldn’t get the stupid thing to read in and assign
the countries correctly. After hours, I finally found a way to make it
work. Needless to say I didn’t get as far as i hoped to get tonight.
Will be up later still hoping to get the cities read in correctly. Will
try to get some of the sorts done as well in the morn.